QLE für Machine funktioniert

+++ src/data/product.py
@@ -14,19 +14,19 @@ class Product:
 
     required_product_type_step_1: ProductionMaterial | None
     processing_step_1: int | None
-    processing_time_step_1: int | None
+    processing_time_step_1: float | None
 
     required_product_type_step_2: ProductionMaterial | None
     processing_step_2: int | None
-    processing_time_step_2: int | None
+    processing_time_step_2: float | None
 
     required_product_type_step_3: ProductionMaterial | None
     processing_step_3: int | None
-    processing_time_step_3: int | None
+    processing_time_step_3: float | None
 
     required_product_type_step_4: ProductionMaterial | None
     processing_step_4: int | None
-    processing_time_step_4: int | None
+    processing_time_step_4: float | None
 
     @property  # only if identification_str is used; one time calculation -> is cached
     def identification_str(self) -> str:
